feat: map resource to asset when receiving RM details
